XC-347: Move to separate module
